feat: Enhance log container and compose management with new UI components and improved functionality
